+++ examples/hello-world/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alphaisyour/example-hello-world
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @alphaisyour/renderer-dom@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes