+++ packages/create-noalphjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-noalph-app
 
+## 0.3.0
+
+### Minor Changes
+
+- v2 upgrade: SSR, file-based routing, image optimization, SEO built-in, CLI animasi, folder .noalph/, icon support
+
 ## 0.2.0
 
 ### Minor Changes