+++ packages/noalph-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @alphaisyour/cli
 
+## 0.3.0
+
+### Minor Changes
+
+- v2 upgrade: SSR, file-based routing, image optimization, SEO built-in, CLI animasi, folder .noalph/, icon support
+
+### Patch Changes
+
+- Updated dependencies
+  - @alphaisyour/shared@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes