+++ packages/noalph-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alphaisyour/core
 
+## 0.3.0
+
+### Minor Changes
+
+- v2 upgrade: SSR, file-based routing, image optimization, SEO built-in, CLI animasi, folder .noalph/, icon support
+
+### Patch Changes
+
+- Updated dependencies
+  - @alphaisyour/shared@0.2.2
+  - @alphaisyour/parser@0.2.2
+  - @alphaisyour/compiler@0.2.2
+  - @alphaisyour/renderer-dom@0.2.2
+  - @alphaisyour/server@0.2.0
+  - @alphaisyour/router@0.2.0
+
 ## 0.2.1
 
 ### Patch Changes