+++ packages/noalph-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @alphaisyour/shared
 
+## 0.2.2
+
+### Patch Changes
+
+- v2 upgrade: SSR, file-based routing, image optimization, SEO built-in, CLI animasi, folder .noalph/, icon support
+
 ## 0.2.1
 
 ### Patch Changes